+++ src/main/java/kr/hs/entrydsm/rollsroyce/domain/admin/presentation/AuthController.java
@@ -1,0 +1,33 @@
+package kr.hs.entrydsm.rollsroyce.domain.admin.presentation;
+
+import kr.hs.entrydsm.rollsroyce.domain.admin.presentation.dto.request.LoginRequest;
+import kr.hs.entrydsm.rollsroyce.domain.admin.service.LoginService;
+import kr.hs.entrydsm.rollsroyce.domain.admin.service.TokenRefreshService;
+import kr.hs.entrydsm.rollsroyce.global.utils.token.dto.TokenResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+
+@RequiredArgsConstructor
+@RequestMapping("/admin/auth")
+@RestController
+public class AuthController {
+
+    private final LoginService loginService;
+    private final TokenRefreshService tokenRefreshService;
+
+    @ResponseStatus(HttpStatus.CREATED)
+    @PostMapping
+    public TokenResponse login(@RequestBody @Valid LoginRequest request) {
+        return loginService.execute(request);
+    }
+
+    @ResponseStatus(HttpStatus.CREATED)
+    @PutMapping
+    public TokenResponse refreshToken(@RequestHeader("X-Refresh-Token") String refreshToken) {
+        return tokenRefreshService.execute(refreshToken);
+    }
+
+}
